+++ CHANGELOG.md
@@ -37,6 +37,7 @@ For details about compatibility between different releases, see the **Commitment
 - `LinkADRReq` MAC command generation for LoRaWAN 1.0 and 1.0.1 end devices.
 - `LinkADRReq` no longer attempts to enable channels which have not yet been negotiated with the end device.
 - Downlink path selection for uplinks which are not LoRa modulated.
+- Console showing deleted collaborator after successful deletion in application collaborator list.
 
 ### Security